+++ include/resources/af.h
@@ -7,10 +7,10 @@ typedef struct af_t {
     unsigned int id;
     unsigned int endurance;
     unsigned int health;
-    int forward_speed;
-    int reverse_speed;
-    int jump_speed;
-    int fall_speed;
+    float forward_speed;
+    float reverse_speed;
+    float jump_speed;
+    float fall_speed;
     af_move moves[70];
     char sound_translation_table[30];
 } af;
